only build the string if it is stale.
yes this may cause oom issues depending on the size, and if that is the
case then an alternative for generating the toString should be found,
because at that size it will take a long time to generate it everytime.

oh, and clean up runNumber issue when not a datadriven test.